% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fks-methods.R
\name{coef.freekt}
\alias{coef.freekt}
\title{Compute Coefficients of B-Splines For Free-Knot Splines}
\usage{
\method{coef}{freekt}(object, ...)
}
\arguments{
\item{object}{An object of class "\code{freekt}" obtained by using one of
the fitting algorithms.}

\item{\dots}{Additional arguments to be passed to coef.freekt.  Currently
ignored.}
}
\value{
A vector containing the coefficients of the B-splines.
}
\description{
This function computes the coefficients of the B-splines for free-knot
splines, given the amount of the penalty (if applicable) and the locations
of the knots.
}
\examples{

x = 0:30/30
truey = x*sin(10*x)
set.seed(10556)
y = truey + rnorm(31, 0, 0.2)
xy.freekt = freelsgen(x, y, degree = 2, numknot = 2, 555)
coef.freekt(xy.freekt)

}
\seealso{
\code{\link{fitted.freekt}} to compute the fitted values and
\code{\link{residuals.freekt}} to compute the residuals.
}
\author{
Steven Spiriti
}
